xero-data-ingestion/app/requirements.txt
Flask==2.3.2
google-cloud-secret-manager==2.15.0
google-cloud-storage==2.11.0
google-cloud-bigquery==3.5.0
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
ratelimit==2.2.1
cachetools==5.3.0
structlog==23.4.0